Move commands and associated handlers into their own files.
