+++ cli/writer.go
@@ -11,23 +11,24 @@ func tabWriter(w io.Writer) *tabwriter.Writer {
 	return tabwriter.NewWriter(w, 0, 0, 1, ' ', tabwriter.AlignRight|tabwriter.Debug)
 }
 
-func printResult(w *tabwriter.Writer, res *response) {
+// PrintResult formats and prints the content of `res` to `w`
+func PrintResult(w *tabwriter.Writer, res *Response) {
 	header := ""
 	spaceRow := ""
-	columnLength := len(res.columns)
+	columnLength := len(res.Columns)
 
-	for i := range res.columns {
+	for i := range res.Columns {
 		tab := " \t "
 		if columnLength == i+1 {
 			tab = " "
 		}
-		header += res.columns[i] + tab
+		header += res.Columns[i] + tab
 		spaceRow += " " + tab
 	}
 
 	fmt.Fprintln(w, header)
 	fmt.Fprintln(w, spaceRow)
-	for _, row := range res.result {
+	for _, row := range res.Result {
 		rowStr := ""
 		for range row {
 			rowStr += "%v \t "
